Gateways now get separate etc dir, second publisher now connects to second gateway

+++ test/container/500-concurrent-publishing/scripts/setup_connected_publisher.sh
@@ -1,10 +1,16 @@
 #!/bin/bash
-#jsudo yum install -y https://cvmrepo.s3.cern.ch/cvmrepo/yum/cvmfs-release-latest.noarch.rpm
-#sudo yum install -y cvmfs cvmfs-server
+
+# Check if the gateway URL is provided as an argument
+if [ -z "$1" ]; then
+  echo "Usage: $0 <CVMFS_GATEWAY_URL>"
+  exit 1
+fi
+
+CVMFS_GATEWAY_URL=$1
+
 yum -y install jq sqlite
 
 FQRN=test.repo.org
-CVMFS_GATEWAY_URL=http://cvmfs-gw1
 CVMFS_STRATUM0_URL=http://cvmfs-s3:9000/mybucket
 CVMFS_SERVER_DEBUG=3 cvmfs_server mkfs -w $CVMFS_STRATUM0_URL/$FQRN \
                          -u gw,/srv/cvmfs/$FQRN/data/txn,$CVMFS_GATEWAY_URL:4929/api/v1 \
